requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
faiss-cpu==1.7.4
PyPDF2==3.0.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
pydantic==2.5.0